+++ src/components/LoanFinding/EmptyState.vue
@@ -1,0 +1,16 @@
+<template>
+	<!-- eslint-disable max-len -->
+	<div class="tw-flex tw-flex-col lg:tw-flex-row tw-gap-2 tw-bg-white tw-px-2 tw-pb-2 lg:tw-py-4 lg:tw-px-8 tw-items-center">
+		<img class="tw-w-8 lg:tw-w-16" src="~@/assets/images/sad_cloud.svg">
+		<h2 class="tw-text-h2">
+			We couldn’t find any loans that match your current filters but here are other recommended loans for you.
+		</h2>
+	</div>
+</template>
+
+<script>
+
+export default {
+	name: 'EmptyState',
+};
+</script>